requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy.orm
dotenv
mysql.connector